Add beaver user to ceph group (if exists)
Adds the beaver user to the ceph group if it exist so beaver can read
the ceph log files. Allows for new groups to easily be added.
Fixed beaver.yml to use the variable name ansible_hostname matching the
ceph playbook that writes out the config file with the name of the log.

Change-Id: Ida517090475c8a980f5d86350089f138980863f8