Fixes for the privatecloud script install
Fixes ops.sh for the new OPSMGR_PRL variable name and way to call the
playbook. Adds a call to target.yml
Fixes privatecloud-mitaka inventory template to match osa-mitaka
Fixes the problem with ruby not being installed on all containers
for nrpe plugins
Fixes ansible errors when there are no compute or storage hosts
Removes the download of the new config.guess, config.sub file
because the server is unreliable. Hardcodes the build parameter
for ppc64le to achieve the same end result.

Change-Id: I2dc9b22f5c41309a6c7c8b7d6cbe748546cc8f06