Fix AIO Builds
This is a companion code drop to ceph-services and os-services
with the same title.  It provides a genesis style inventory for
AIO, so that opsmgr can utilize a common algorthim for its
playbooks.  There is very little genesis based information in the
playbooks but they assume the file exists.

Change-Id: I06ad1dc074ead4f63b2bcfdc3202c86553e30ba9
(cherry picked from commit 86b71aeb81dcd9ef0574565c144f546467a86ee6)