Updates to install opsmgr on Power + Ubuntu 16.04
For both elasticsearch and logstash:
jre-7 is required but not shipped with ubuntu 16.04. Added the debain
experimental repo to install jre-7.
Installed elasticsearch and logstash using the .deb package for both
ppc64le and x86 to follow a common code path.
Patched the elasticsearch plugin shell script to remove the '-client'
java option that is not valid for the ppc64le jre.
Ensured /var/lib/elasticsearch/openstack was created with the proper
permissions to avoid an issue starting elasticsearch
For logstash handled the case when ansible_processor_vcpus is 0
because of bug in ansible for ppc64le

For Kibana:
Kibana doesn't provide a .deb package for ppc64le. Kibana is written
in nodejs. So we download the kibana x86 tar.gz file and replace
the nodejs binary with the ppc64le nodejs binary. The install path
for x86 installs was not modified. A new install path for ppc64le
was written.

For Nagios:
Upgraded to latest version, 4.2.1 to 4.2.2
Updated required packages for Ubuntu 16.04
Compiled ppc64le version with the build tag ppc64le-unknown-linux-gnu
Added a Nagios service definition for Ubuntu 16.04 to fix a error with
the nagios service not being found.

Change-Id: If5106301418bf59ae0bc3581309a8e4c83f870c2