Fixes for integration testing of Trove
Added the genesis inventory and floating ip to the list of blacklisted
ips when creating containers.
Rewrote beaver force stop to not kill all processes with 'beaver' in
the string. ps -ef on the host returns all processes running across
all containers, so service beaver stop on the host was killing beaver
across all containers.
kibana needs to reference the elasticsearch floating ip to be HA.
Added retries and delay to the creation of the opsmgr sql user.

Change-Id: I0128f0848132697d45e781bc6d34ebf108315463