Kibana: Import logstash index with timestamp field
This change was integrated before but was lost during the big
recipe change. Redoing this change.
Currently after deploying ELK, the first time you sign into elk you
need to select a field to be the timestamp field used for all
visualizations. This logic does that automatically, so when the user
logs in, they are taken straight to the dashboard.

Also this change adds the default logstash template that tells
elasticsearch to create .raw fields for strings. Something the
newest version of logstash doesn't always seem to do.

Change-Id: Ie91dc1fa15cdf05f38dbb00749d1a2385536965b