Kibana: Import logstash index with timestamp field
With the change in branch-v1 to import the logstash template, the
kibana UI loads with just a title bar, clicking on any option
gives an error that says you must set the default index but the
screen to set the index doesn't display.
This patches sets that default index specifies the timestamp
field to use.

Change-Id: I20fe713bd249e424f17aeea09737c592aaa45f56