Fix database permissions during install
The mysql user ansible creates a user with the priv specified
it is unable to loop through a list of items and do multiple privilages
as you only end up with the last one. So we wildcard the permissions
to be any database that starts with opsmgr.
We delegate the task to the first host to avoid a race condition.

Change-Id: I18740f0c3e6b210e1063a2070013491d59925c37