Fix sshd_config to allow PasswordAuthentication
If PasswordAuthentication is set to no it will fail the ssh login
This change updates PasswordAuthentication to yes to fix this issue
Also comment out the interval settings avoiding ssh timeouts

removed the state: absent from the updatessh.yml - the documentation
wasn't clear. I was attempting to ensure the ServerAliveInterval was
not present before attempting to add this line to the ssh_config
file. With absent added the code appeared to work properly, but may
be extra.
Change-Id: I0902d889d12c2bdb1e58656e3afba676b0fa120a