Add retry to the step: enable the nagios service
Recently the step to enable the nagios service to start on reboot has
been failing. The enable command appears to work, but prints out a
warning message and returns a non zero code. On the next run of the
script, ansible sees the service is already enabled and doesn't try to
enable it. Adding retry avoids this problem for the reason above.
This should be investigated deeper when more time allows.

Change-Id: I26e104b0d701ab62c85fffa293302581d0ccd33b