Fix for 3 controller install: Setup Hosts
Without this fix, the steps to setup the ssh key for the container
was done on the localhost. This didn't work in the case where the
containers were on 3 different controllers. Moved this logic into
the setup of the containers. It still depends on the rootfs, but
other parts of the code have the same dependency.

Change-Id: Ic02d58361e493b5044cc4b6ffc9d1b69e3f36ee7