Fix undefined var deploying targets to ceph_osd
When the ceph_osd provisioning plugin is running, the update.xml
file is called to update the nagios_target with additional checks
for ceph. This didnd't contain the logic to look for vars.yml
in the same directory. Instead of using vars.yml, moved the
default values to the nagios_target/defaults/main.yml file
and updated update.yml to look for a variable file defined
by the provisioning plugin

Change-Id: I6311047f07874648264fd01a4e95fb0770a1cd4c