add patch package to nagios agent
Last end to end test failed because ceph-osd and swift nodes
didn't have the patch command installed. This ensures the
patch command is installed prior to patching nrpe.

Also replaced ignore_errors:true with failed_when:false
in targets playbook so not to alarm users when there are
no provisioning plugins for an inventory type.

Change-Id: Ia7471187d097d3c133bd1b5fdad18c2d37de3a3b