Fix hang by using a dynamic include
The opsmgr install was hanging when running the tasks:
'insert common checks into nagios cmd files' against all containers.
It seems this combination of a lot of hosts, and two long list used
in with_nested was causing ansible to hang. Since for all containers
the when: has_agent == true statement is always false, I reworked
the playbook to make the has_agent check a dynamic include.

static: no, on the include tasks is required to do this, that
logic can probably be copied into other playbooks if for nothing
else then to reduce the amount of logging for tasks skipped.

Change-Id: Ib94aa0ddd820a65786fcda77b24c0c07f8704622