Remove ignore_error statement when using redirect.yml
redirect.yml copies a playbook to /tmp on a host and runs
it on that host, since it had ignore_errors: true if the
playbook failed the user was not notifed.

In this case the playbook failed downloading nrpe.
I moved the until check to be just below register like the examples
online. I haven't had a failure since making this change.

Change-Id: Id1789fe6a185642d702b2da1a5dce25d17e588af