Add checksum to nrpe download
Discovered that even when the download failed, the task would succeed
because it downloaded a file called mirrorproblem. The next task would then
fail because the expected file didn't exist. Added checksum to the download
so get_url will fail and retry if the download doesn't match the expected
checksum.

Change-Id: I90ad08e301766b6d284ff270511ea09cb0038abd