Changes to customize OpsMgr standalone minimal OSA installation
Two changes:
- remove bridges br-vlan, br-vxlan and br-storage + configure /etc/network/interface.d files
- rename lxc group name on OSA minimal install containers to denote "opsmgr" instead of "openstack"

We use OSA's AIO setup in OpsMgr standalone and it hardcodes creation of the bridges.
So it's not enough to only change the user config yml for an AIO installation.
Instead of patching up OSA's code in several different places I've decided the approach to remove after run was easier.
Same thing for the configuration of the lxc container groups - changing and restarting the containers is easier.

Change-Id: I19c737e3eab82c133522d31766fdca790558fb9f