Integration Test Fixes
The main fix here is to read the reference-architecture and if
set to ceph-standalone, install the nagios monitors for ceph-mon.
Since the hypen is invalid in variable names in ansible, used
perl to parse the yml and return the list of reference-architectures.

Add lsvpd package to all systems so it can find the lscfg command
required to get machine type model and serial number of bare metal
systems
Monitor the br-replication interface
Commented out the ceph monitor check for br-storage which doesn't
exist in ceph-standalone

Change-Id: I8742c28ce3a63f3f38be2eafb4b583b49e49e859