Remove hardcoded path to swift venv
Remove the hardcoded path to bin directory in the swift venv.
Instead resolve the path in the OSA recipes and save this as a
variable swift_bin. Reference this variable in a template py file
that the python plugins can import.
For the one shell nagios plugin, treat it as a template file as
replace the swift_bin path during the install.

Change-Id: I61224870ebd119665d491eb54f2c3e16c5ed69d8