Run swift-dispersion-populate post install
The swift-dispersion-populate tool makes up random container and
object names until they fall on distinct partitions.
Nagios will then call swift-dispersion-report to check the health
of each of these containers and objects for the life of the cluster.

This fix elimantes the error in nagios that the swift-dispersion-populate
command has not been run yet.

Change-Id: If2adc1bf7ab422c3862cc309da237cd8a456a274