Filebeat: Fix ceph monitoring in converged stack
When running ceph-monitor and osa-newton on the same host, filebeat
was crashing as soon as it was starting. Debugging this I figured
out it was because entries to monitor the ceph log files were in
both the OSA-Newton config file and the ceph config files. Having
the duplicate entiries caused filebeat to crash.

I also noticed the ceph dashboard was incomplete, as it appeared
kibana was including logs with no values in the graphs that averaged
things like bytes read. I fixed the visualizations to only include
logs entries where the value has been set and this fixed the problem.

Change-Id: I674384bf2b8d8fefbdccb665ecce791561521222