Metricbeat: Import template before importing index
Importing the metricbeat index fixed the problem with kibana
not formatting the fields as bytes and displaying human
readable fields. However it introduced a new error that
some fields were not analyzed. Metricbeat will automatically
import the template to resolve this problem, but it can't
import the template if the index has already been created.
So we know import the template and then import the index.

The template files here were extracted from the metricbeat
install and have not been modified. Including the templates
for both elasticsearch 2.x and 5.x

Change-Id: Ia29f4e4d8c61371ca1f79d2c745696b74e285c82