Update opsmgr (elk) monitoring for minimal install
Updated the new opsmgr monitoring for elasticsearch, logstash and
kibana to handle the case for the minimal install where these services
are not installed in LXC containers.

Ran into two issues that affect the main code path:

1) eth0 and eth1 monitoring is valid for LXC ocntainers but not for
when installing on the host directly. Removed eth0 and eth1 monitoring
will look at something different for network monitoring in the future.

2) With the install on the host directly the process count for
elastic and all was higher then in LXC. It seemed to be including the
grep command check. Increased the check from no more then 1 process to
no more then 3 processes.

Change-Id: I8ffa0fc754ae731a621f462d69e2c91aee717ae0