Added clean scripts for the standalone recipe
Added clean-prereq.sh to run the documented commands to uninstall
OpenStack Ansible all in one install.
Added clean-opsmgr.sh to run the clean.yml playbook to remove
opsmgr containers.

Added a patch for the OpenStack Ansible all in one install to prevent
it from changing the existing authentication settings in
/etc/ssh/sshd_config

Makes fixes to ensure the patches were applied before running the OSA
bootstrap scripts and it didn't apply them a second time because that
would fail.

Change-Id: Ifbc51f3cbc3181aef87f7ee0d6cde2c4b50d4951