Patch OSA to remove sshd_config idle timeouts
The OSA installs sets an idle timeout of 15 minutes for ssh.
This causes scripts that are running via ansible ssh commands to
timeout. Before we updated this value after the opsmgr install was
done, but now we are patching the OSA code so it never sets this value.

The updatessh.yml playbook will be deleted when the osa-newton recipe
is deleted. (Refactoring on the todo list)

Change-Id: Ic9c0ee140d4b35cbc151f3f81d7ff355b8b61655