Changes to fix standalone installation and cleanup.
In a standalone installation the customer doesn't use root.
Also genesis doesn't automatically set passwordless ssh/sudo.
There's also need to perform a clean of the environment if
a reinstall is done. So, the following changes are needed:
- all scripts: wrap code before/after to use genesis key in .ssh
- all "prereq" scripts: osa needs the same for the root user
(OSA hardcodes using the root id_rsa.pub key in lxc image)
- nrpe.cfg needs "routing_address" added to allowed_hosts
- db_password.yml needs to use a location normal users access
- on a reinstall creation of .ssh/config must happen without
access to actual remote hosts (otherwise setup will fail)
I've moved the logic to create config out of "hosts.yml", so
now running "setup.yml" is enough to re-set the .ssh/config,
even in a re-install scenario (so it can be cleaned up).

Change-Id: Ic30d1c017311eb2e9f17f9cd9782fb47c3608259