Add checksum and retry logic for all downloads
Updates for branch-v3 with older levels of the elk stack.

Added checksums for every get_url call the normal code path uses.
Added retry_logic for when the download fails or the checksums
don't match.
Since unarchive being used to download a file doesn't support
checksums, switched to using get_url to download the file and unarchive
to extract the downloaded file.

Change-Id: I3a960d630ace3100b96ca1b4aa652d50302d33e0