Fixes for acclerated db testing
ifup was causing the network to come up by failing ansible
before it returned an error RNETLINK file exists. To workaround
that problem we decided to ignore this error until more investigation
can be done.

When running clean-opsmgr.sh we should also be removing the old data
from the system, so we remove the directory the containers mount
after the containers have been removed.

Change-Id: Ib81c3a35597bb101f8d9bcc67f2f2d5530ba8428