Elasticsearch Container Bug
The install of elastic search tries to modify kernel parameters which
in an LXC container are read-only. This doesn't fail the install
because it is a post install step, and on retry ansible finds the
package is already installed, but it does fail the next attempt at
using apt install, because it retries the configuration of
elasticsearch. To avoid this problem we set the environment variable
ES_SKIP_SET_KERNEL_PARAMETERS to true

Change-Id: I7460d2de9b674a22c55359ce4da7264db314299a