clone beats version 5.2.2
The git clone task didn't specify a version, so git clone was
always downloading the master branch which can contain breaking
changes at any time. We need to force the git clone to use
the tag v5.2.2 for metricbeat and filebeat

Also we have a problem during the install when running the base
playbook against all targets and a single node had two different
hostnames. One thread would delete the directory, the other thread
would fail to delete the directory because it already had been
deleted. Added retry for a workaround to this problem.

Change-Id: Ifab1d438109625a6a19c282029f3b8ce93b9efbf