Disable loopback for standalone install
Loopback devices were being created for cinder, swift and nova
that were never being used. Those are now disabled by setting
a boolean variable for each one to no (or false)

In addition, modified the minimum install logic so instead of removing
the unncessary yaml files after bootstrap-prereq runs, we are
modifying the input to bootstrap-prereq to specify a scenerio that
only installs horizon and keystone.

Change-Id: I9af0dbe3180d32527dfdf063114f4a1d40358dc1