Remove elasticsearch plugins from ELK 2.4
In the master branch we removed the elasticsearch plugins when
moving to ELK 5.X, as we have never used them and they
are not used in rpc openstack.

Because of the need to download everything thru the web proxy, we
either had to change how we are calling the plugin install script
or not install any plugins. For this reason they are being removed
now.

Change-Id: Ib995c36fe197e9bfb6a9f07cf9ea48a5bd0b9672