Fixes for deployment variables
Renamed get_env_vars.py to set_env_vars.py and re-wrote the logic
to have the python script write out the variables files based
on the deployment environments variables stored in the genesis
inventory file. This was to fix a issues with leading spaces getting
removed.

Added environment variables to no_proxy container ips for a few
cases where curl http:// is used to invoke an elasticsearch or
kibana api call

Change-Id: I9f7b79ce7adc580447c8285b68afb0bd8030156e