Avoid hardcoding path to localsettings.py
In a juju installed environment local_settings.py is not under
/etc/horizon. Researching the horizon code, they find local_settings
by looking where the django python code is installed and adding
'local/local_settings.py' to that directory. In an ansible environment
local/local_settings.py is a symlink to /etc/horizon/local_settings.py
so this logic should work for all installs of horizon.

Also set the can_register method for the panel to accept keystone
version 2.
Set the PBR_VERSION in the python setup.py file so it doesn't need to
be set when called via the juju charmhelper pip install method.

Change-Id: Ib483ffd1c1adc306e14ccb8f6a48130b2266e4c5